Behavioral and Visual Changes:
Toolbar Positioning:
When the application language is set to English (LTR), the editor's floating toolbar will be positioned on the left side of the canvas.
When the language is set to Arabic (RTL), the toolbar will remain in its current position on the right side of the canvas.

Popup Panel Positioning:
The "Zoom Tool" and "Mask Color/Brush Size" popup panels will now open adjacent to the toolbar, adapting to its new position.
In English, the popups will open to the right of the left-aligned toolbar.
In Arabic, the popups will open to the left of the right-aligned toolbar.