Here are the changes I've made:
New "Studio" Section: I've added a new "Studio" content type to the landing page. It appears as the first tab and, like the "Video" and "Audio" sections, is marked as "Coming Soon." I also created a new icon to represent it.

Translation Correction: I've updated the Arabic translation file, changing all instances of "استوديو" to "أستوديو" to use the correct hamza.

Version Number Update:
The version number in the page footer has been changed to v3.6.7